extract azpsconfig class
