+++ src/PowerShell/AzPSLogin.ts
@@ -1,11 +1,10 @@
 import * as core from '@actions/core';
 import * as exec from '@actions/exec';
 import * as io from '@actions/io';
-import * as os from 'os';
-import * as path from 'path';
 
 import AzPSScriptBuilder from './AzPSScriptBuilder';
 import AzPSConstants from './AzPSConstants';
+import AzPSConfig from './AzPSConfig';
 import { LoginConfig } from '../common/LoginConfig';
 
 interface PSResultType {
@@ -23,8 +22,8 @@ export class AzPSLogin {
 
     async login() {
         core.info(`Running Azure PowerShell Login.`);
-        setPSModulePathForGitHubRunner();
-        await importLatestAzAccounts();
+        AzPSConfig.setPSModulePathForGitHubRunner();
+        await AzPSConfig.importLatestAzAccounts();
         const [loginMethod, loginScript] = await AzPSScriptBuilder.getAzPSLoginScript(this.loginConfig);
         core.info(`Attempting Azure PowerShell login by using ${loginMethod}...`);
         core.debug(`Azure PowerShell Login Script: ${loginScript}`);
@@ -63,36 +62,4 @@ export class AzPSLogin {
         }
         return result.Result;
     }
-}
-
-export function setPSModulePathForGitHubRunner() {
-    const runner: string = process.env.RUNNER_OS || os.type();
-    switch (runner.toLowerCase()) {
-        case "linux":
-            pushPSModulePath(AzPSConstants.DEFAULT_AZ_PATH_ON_LINUX);
-            break;
-        case "windows":
-        case "windows_nt":
-            pushPSModulePath(AzPSConstants.DEFAULT_AZ_PATH_ON_WINDOWS);
-            break;
-        case "macos":
-        case "darwin":
-            core.warning(`Skip setting the default PowerShell module path for OS ${runner.toLowerCase()}.`);
-            break;
-        default:
-            core.warning(`Skip setting the default PowerShell module path for unknown OS ${runner.toLowerCase()}.`);
-            break;
-    }
-}
-
-async function pushPSModulePath(psModulePath: string) {
-    process.env.PSModulePath = `${psModulePath}${path.delimiter}${process.env.PSModulePath}`;
-    core.debug(`Set PSModulePath as ${process.env.PSModulePath}`);
-}
-
-export async function importLatestAzAccounts() {
-    let importLatestAccountsScript: string = AzPSScriptBuilder.getImportLatestModuleScript(AzPSConstants.AzAccounts);
-    core.debug(`The script to import the latest Az.Accounts: ${importLatestAccountsScript}`);
-    let azAccountsPath: string = await AzPSLogin.runPSScript(importLatestAccountsScript);
-    core.debug(`The latest Az.Accounts used: ${azAccountsPath}`);
 }
