+++ src/cleanup.ts
@@ -1,6 +1,5 @@
 import * as core from '@actions/core';
-import { setUserAgent } from './common/Utils'; 
-import { cleanupAzCLIAccounts, cleanupAzPSAccounts } from './common/Utils';
+import { setUserAgent, cleanupAzCLIAccounts, cleanupAzPSAccounts } from './common/Utils'; 
 
 async function cleanup() {
     try {
@@ -11,7 +10,7 @@ async function cleanup() {
         }
     }
     catch (error) {
-        core.setFailed(`Login cleanup failed with ${error}.`);
+        core.setFailed(`Login cleanup failed with ${error}. Make sure 'az' is installed on the runner. If 'enable-AzPSSession' is true, make sure 'pwsh' is installed on the runner together with Azure PowerShell module.`);
         core.debug(error.stack);
     }
 }
